# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LA RC Hub Communication Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Announcement {
  id: number;
  title: string;
  content: string;
  author: string;
  pinned: boolean;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface AnnouncementInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  /** @minLength 1 */
  author: string;
  pinned?: boolean;
}

export interface AnnouncementUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  content?: string;
  pinned?: boolean;
}

export interface Channel {
  id: number;
  name: string;
  description: string;
  messageCount?: number;
  createdAt: string;
}

export interface ChannelInput {
  /** @minLength 1 */
  name: string;
  description: string;
}

export interface Message {
  id: number;
  channelId: number;
  content: string;
  author: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
  /** @minLength 1 */
  author: string;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  location: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  organizer: string;
  createdAt: string;
}

export interface EventInput {
  /** @minLength 1 */
  title: string;
  description: string;
  location: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  organizer: string;
}

export interface EventUpdate {
  title?: string;
  description?: string;
  location?: string;
  startDate?: string;
  /** @nullable */
  endDate?: string | null;
}

export interface Member {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  joinedAt: string;
}

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  role: string;
  bio?: string;
  avatarUrl?: string;
}

export interface MemberUpdate {
  name?: string;
  role?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface HubStats {
  totalMembers: number;
  totalAnnouncements: number;
  totalChannels: number;
  totalMessages: number;
  totalEvents: number;
  upcomingEventsCount: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  actor: string;
  occurredAt: string;
}