# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LA RC Hub Communication Hub API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: announcements
    description: Community announcements
  - name: channels
    description: Communication channels
  - name: messages
    description: Channel messages
  - name: events
    description: Community events
  - name: members
    description: Hub members
  - name: stats
    description: Hub statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List all announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create a new announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Get an announcement by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List all channels
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"

  /channels/{id}:
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: Delete a channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /channels/{channelId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Post a message to a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all members
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Add a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

  /members/{id}:
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Remove a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get hub statistics summary
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubStats"

  /stats/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [stats]
      summary: Get recent activity feed
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /stats/upcoming-events:
    get:
      operationId: getUpcomingEvents
      tags: [stats]
      summary: Get upcoming events (next 5)
      responses:
        "200":
          description: Upcoming events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Announcement:
      type: object
      required: [id, title, content, author, pinned, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        author:
          type: string
        pinned:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time

    AnnouncementInput:
      type: object
      required: [title, content, author]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        author:
          type: string
          minLength: 1
        pinned:
          type: boolean

    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        pinned:
          type: boolean

    Channel:
      type: object
      required: [id, name, description, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    ChannelInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    Message:
      type: object
      required: [id, channelId, content, author, createdAt]
      properties:
        id:
          type: integer
        channelId:
          type: integer
        content:
          type: string
        author:
          type: string
        createdAt:
          type: string
          format: date-time

    MessageInput:
      type: object
      required: [content, author]
      properties:
        content:
          type: string
          minLength: 1
        author:
          type: string
          minLength: 1

    Event:
      type: object
      required: [id, title, description, location, startDate, organizer, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        organizer:
          type: string
        createdAt:
          type: string
          format: date-time

    EventInput:
      type: object
      required: [title, description, location, startDate, organizer]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        organizer:
          type: string

    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time

    Member:
      type: object
      required: [id, name, role, joinedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        joinedAt:
          type: string
          format: date-time

    MemberInput:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string

    MemberUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string

    HubStats:
      type: object
      required: [totalMembers, totalAnnouncements, totalChannels, totalMessages, totalEvents, upcomingEventsCount]
      properties:
        totalMembers:
          type: integer
        totalAnnouncements:
          type: integer
        totalChannels:
          type: integer
        totalMessages:
          type: integer
        totalEvents:
          type: integer
        upcomingEventsCount:
          type: integer

    ActivityItem:
      type: object
      required: [id, type, description, actor, occurredAt]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        actor:
          type: string
        occurredAt:
          type: string
          format: date-time