# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LA RC Hub Communication Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  actor: string;
  occurredAt: Date;
}