# lib/api-zod/src/generated/types/announcementUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LA RC Hub Communication Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface AnnouncementUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  content?: string;
  pinned?: boolean;
}