# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LA RC Hub Communication Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Channel {
  id: number;
  name: string;
  description: string;
  messageCount?: number;
  createdAt: Date;
}