# lib/api-zod/src/generated/types/eventInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LA RC Hub Communication Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface EventInput {
  /** @minLength 1 */
  title: string;
  description: string;
  location: string;
  startDate: Date;
  /** @nullable */
  endDate?: Date | null;
  organizer: string;
}