# lib/api-zod/src/generated/types/hubStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LA RC Hub Communication Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface HubStats {
  totalMembers: number;
  totalAnnouncements: number;
  totalChannels: number;
  totalMessages: number;
  totalEvents: number;
  upcomingEventsCount: number;
}