# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LA RC Hub Communication Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  channelId: number;
  content: string;
  author: string;
  createdAt: Date;
}