# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LA RC Hub Communication Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  /** @minLength 1 */
  content: string;
  /** @minLength 1 */
  author: string;
}